Refactor cloud function into multiple files part 7 - the rest
