+++ examples/cloud-operations/network-dashboard/cloud-function/metrics/instances.py
@@ -1,9 +1,5 @@
 from code import interact
 from collections import defaultdict
-from google.protobuf import field_mask_pb2
-from google.cloud import asset_v1
-#TODO: Workaround
-import main as main
 from . import metrics, networks, limits
 
 def get_gce_instance_dict(config: dict):
@@ -47,23 +43,18 @@ def get_gce_instances_data(config, metrics_dict, gce_instance_dict, limit_dict):
       Returns:
         gce_instance_dict
   '''
-  # TODO: move out
-  # Existing GCP Monitoring metrics for GCE instances
-  metric_instances_limit = "compute.googleapis.com/quota/instances_per_vpc_network/limit"
 
   for project in config["monitored_projects"]:
     network_dict = networks.get_networks(config, project)
 
-    #TODO
     current_quota_limit = limits.get_quota_current_limit(config, f"projects/{project}",
-                                                  metric_instances_limit)
+                                                  config["limit_names"]["GCE_INSTANCES"])
     if current_quota_limit is None:
       print(
           f"Could not write number of instances for projects/{project} due to missing quotas"
       )
 
-    # TODO: workaround
-    current_quota_limit_view = main.customize_quota_view(current_quota_limit)
+    current_quota_limit_view = metrics.customize_quota_view(current_quota_limit)
 
     for net in network_dict:
       limits.set_limits(net, current_quota_limit_view, limit_dict)
